qcacmn: Modify 11BE Opclass changes
New Opclass for 11BE mode is introduced. The new opclass for 11BE is
included in global opclass according to 802.11_2016-SPEC.

Another change is modify the type of chan_spacing variable in
struct reg_dmn_op_class_map_t from uint8_t to uint16_t. The places
where the chan_spacing variable used is also modified with the updated
data type.

CRs-Fixed: 2897863

+++ umac/regulatory/core/src/reg_opclass.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2020 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2017-2021 The Linux Foundation. All rights reserved.
  *
  *
  * Permission to use, copy, modify, and/or distribute this software for
@@ -59,7 +59,7 @@ uint16_t reg_dmn_get_opclass_from_channel(uint8_t *country, uint8_t channel,
  */
 uint8_t reg_dmn_get_opclass_from_freq_width(uint8_t *country,
 					    qdf_freq_t freq,
-					    uint8_t ch_width,
+					    uint16_t ch_width,
 					    uint16_t behav_limit);
 
 /**
@@ -308,7 +308,7 @@ static inline uint16_t reg_dmn_get_opclass_from_channel(
 static inline
 uint8_t reg_dmn_get_opclass_from_freq_width(uint8_t *country,
 					    qdf_freq_t freq,
-					    uint8_t ch_width,
+					    uint16_t ch_width,
 					    uint16_t behav_limit)
 {
 	return 0;
