+++ umac/regulatory/dispatcher/src/wlan_reg_services_api.c
@@ -218,7 +218,7 @@ uint8_t wlan_reg_get_band_cap_from_op_class(const uint8_t *country,
 
 uint8_t wlan_reg_get_opclass_from_freq_width(uint8_t *country,
 					     qdf_freq_t freq,
-					     uint8_t ch_width,
+					     uint16_t ch_width,
 					     uint16_t behav_limit)
 {
 	return reg_dmn_get_opclass_from_freq_width(country, freq, ch_width,
